organize tables in one place
