+++ src/components/Tables/TopClaimers/index.module.scss
