+++ src/pages/UniswapProtocolPage/UniswapProtocolPage.js
@@ -4,7 +4,7 @@ import { Helmet } from 'react-helmet'
 import { UNISWAP_METRIC_BOUNDARIES_QUERY, useRestrictedInfo } from './hooks'
 import ClaimersWidgets from '../../components/ClaimersWidgets'
 import { ChartWidget } from '../../components/ClaimersWidgets'
-import TopClaimersTable from '../../components/ClaimersWidgets/TopClaimers/TopClaimersTable'
+import TopClaimersTable from '../../components/Tables/TopClaimers'
 import TopExchangesTable from '../../components/Tables/TopExchanges'
 import { DesktopOnly } from '../../components/Responsive'
 import { Block } from '../StablecoinsPage/StablecoinsPageStructure'
